requirements.txt
openai
langchain
faiss-cpu
pandas
python-dotenv
tiktoken
sentence-transformers
qrcode[pil]
SpeechRecognition
pydub
gTTs
streamlit
streamlit-audiorec
flask
twilio
pygame
rasa
pyyaml
numpy
scikit-learn